main impl of lazy iterators
