+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 byteorder = "1.5.0"
 downcast-rs = "1.2.1"
+dyn-clone = "1.0.17"
 enum-iterator = "2.0.0"
 lazy_static = "1.4.0"
 rustc-hash = "1.1.0"
